% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_called}
\alias{function_called}
\title{Checks if a function is called optionally with certains args}
\usage{
function_called(code, fn, index = 1, has_args = NULL, arg_values = NULL)
}
\arguments{
\item{code}{quosure list of input code}

\item{fn}{name of function to search for}

\item{index}{which instance of the call to the function should we
look for the args}

\item{has_args}{character vector of args to check that are called
in the function call}

\item{arg_values}{named list of character vectors of the exact arguments
values some arguments should take}
}
\value{
logical, `TRUE` if all conditions are met, `FALSE` otherwise
}
\description{
Checks if a function is called in code
}
